+++ doc/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 v0.3 (*unreleased*)
 -------------------
 - add diff and color diff modes (:pull:`56`)
+- support `black`'s string normalization option (:pull:`59`)
 
 
 v0.2 (01 October 2020)